+++ packages/g-canvaskit/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @antv/g-canvaskit
 
+## 0.10.27-next.3
+
+### Patch Changes
+
+-   e4c1645ae: Test.
+-   Updated dependencies [e4c1645ae]
+    -   @antv/g-plugin-canvas-path-generator@1.3.22-next.3
+    -   @antv/g-plugin-canvaskit-renderer@1.4.0-next.3
+    -   @antv/g-plugin-dom-interaction@1.9.22-next.3
+    -   @antv/g-plugin-canvas-picker@1.10.24-next.3
+    -   @antv/g-plugin-html-renderer@1.9.25-next.3
+    -   @antv/g-plugin-image-loader@1.4.0-next.3
+    -   @antv/g-lite@1.3.0-next.3
+
 ## 0.10.27-next.2
 
 ### Patch Changes