+++ packages/g-mobile-svg/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @antv/g-mobile-svg
 
+## 0.10.27-next.3
+
+### Patch Changes
+
+-   e4c1645ae: Test.
+-   Updated dependencies [e4c1645ae]
+    -   @antv/g-plugin-mobile-interaction@0.9.22-next.3
+    -   @antv/g-plugin-svg-renderer@1.11.0-next.3
+    -   @antv/g-plugin-svg-picker@1.9.26-next.3
+    -   @antv/g-plugin-dragndrop@1.8.22-next.3
+    -   @antv/g-plugin-gesture@1.2.10-next.3
+    -   @antv/g-lite@1.3.0-next.3
+
 ## 0.10.27-next.2
 
 ### Patch Changes