+++ packages/g-plugin-canvas-picker/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @antv/g-plugin-canvas-picker
 
+## 1.10.24-next.3
+
+### Patch Changes
+
+-   e4c1645ae: Test.
+-   Updated dependencies [e4c1645ae]
+    -   @antv/g-plugin-canvas-path-generator@1.3.22-next.3
+    -   @antv/g-plugin-canvas-renderer@1.10.0-next.3
+    -   @antv/g-lite@1.3.0-next.3
+
 ## 1.10.24-next.2
 
 ### Patch Changes