+++ packages/g-plugin-device-renderer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @antv/g-plugin-device-renderer
 
+## 1.10.0-next.3
+
+### Patch Changes
+
+-   e4c1645ae: Test.
+-   Updated dependencies [e4c1645ae]
+    -   @antv/g-plugin-image-loader@1.4.0-next.3
+    -   @antv/g-lite@1.3.0-next.3
+
 ## 1.10.0-next.2
 
 ### Patch Changes