+++ packages/g-plugin-svg-renderer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/g-plugin-svg-renderer
 
+## 1.11.0-next.3
+
+### Patch Changes
+
+-   e4c1645ae: Test.
+-   Updated dependencies [e4c1645ae]
+    -   @antv/g-lite@1.3.0-next.3
+
 ## 1.11.0-next.2
 
 ### Patch Changes