+++ packages/g-plugin-zdog-canvas-renderer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @antv/g-plugin-zdog-canvas-renderer
 
+## 1.2.27-next.3
+
+### Patch Changes
+
+-   e4c1645ae: Test.
+-   Updated dependencies [e4c1645ae]
+    -   @antv/g-canvas@1.11.27-next.3
+    -   @antv/g-lite@1.3.0-next.3
+
 ## 1.2.27-next.2
 
 ### Patch Changes