+++ packages/g-plugin-zdog-svg-renderer/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @antv/g-plugin-zdog-svg-renderer
 
+## 1.2.27-next.3
+
+### Patch Changes
+
+-   e4c1645ae: Test.
+-   Updated dependencies [e4c1645ae]
+    -   @antv/g-plugin-svg-renderer@1.11.0-next.3
+    -   @antv/g-lite@1.3.0-next.3
+    -   @antv/g-svg@1.10.26-next.3
+
 ## 1.2.27-next.2
 
 ### Patch Changes